+++ app/src/main/java/com/example/mingle/domain/Common.java
@@ -11,7 +11,7 @@ public abstract class Common {
     public abstract String getMusicUri();
     public abstract String getTitle();
     public abstract String getArtist();
-    public abstract int getAlbum_id();
+    public abstract long getAlbum_id();
     public abstract String getAlbumImgUri();
     //public abstract Bitmap getAlbumImgBitmap(Context context, int album_id);
     public abstract Bitmap getAlbumImgBitmap();
